Актуализировать исходный код из seminar1_advanced_final
Аналогично базовому треку: ветка seminar2_advanced_review была создана
из более старой версии файлов доменной модели — часть кода откатывала
уже принятые и запушенные в seminar1_advanced_final правки. Семинар 2
не затрагивает эти файлы по плану (кроме Stock — см. ниже):

- Instrument.java — восстановлен NOTICE-комментарий про инвариант
  тикера (был потерян).
- Bond.java — couponRate обратно BigDecimal (был откачен до double —
  прямое нарушение жёсткого правила курса «деньги = BigDecimal, не
  double»; тот же баг, что чинили после ревью семинара 1). Заодно
  поправлен BondFactory.java (единственный новый файл семинара 2,
  который создавал Bond со старой double-сигнатурой) — передаёт
  BigDecimal.ZERO вместо 0.0.
- Etf.java — убран внесённый заново тривиальный class-Javadoc.
- Portfolio.java / PortfolioImpl.java — восстановлены как интерфейс без
  спойлеров + array-based реализация (были заменены на конкретный класс
  с List/ArrayList/Collections.unmodifiableList — коллекции по плану
  курса проходятся только в семинаре 5; PortfolioImpl был полностью
  удалён). Portfolio семинаром 2 не используется ни в одном файле,
  откат безопасен.
- Quote.java — восстановлен switch по changePercent.signum() с явным
  случаем ноля (▬), был откачен на тернарник без различения нуля (тот же
  баг, что чинили после ревью семинара 1); убрана устаревшая ссылка на
  несуществующий «этап 4.7» в Javadoc.
- JolDemo.java / MemoryDemo.java — убраны зашитые в Javadoc номера
  шагов ("Этапа 6"), уже убранные в seminar1_advanced_final по той же
  причине (см. отдельный коммит после ревью PR #32). MemoryDemo.java
  дополнительно: возвращено ожидание Enter через Scanner вместо
  Thread.sleep(10_000) — тот же баг, что чинили после ревью семинара 1
  ("MemoryDemo переведён с Thread.sleep на ожидание ввода").
- Stock.java — не откат, а слияние: сохранено новое "implements
  DividendPaying" (правка семинара 2, этап 1), но восстановлено деление
  с явным RoundingMode.HALF_UP и NOTICE-комментарий про BigDecimal;
  убрана ссылка на несуществующий «этап 4.6».

Проверено: полная сборка (advanced/src, вкл. JOL/cglib classpath)
компилируется, CLI-команды add/list прогнаны вживую.

+++ advanced/src/main/java/academy/backend/market_pulse/model/Bond.java
@@ -1,21 +1,20 @@
 package academy.backend.market_pulse.model;
 
-/**
- * Облигация.
- */
+import java.math.BigDecimal;
+
 public class Bond extends Instrument {
 
-    private final double couponRate;
+    private final BigDecimal couponRate;
     private final int maturityYear;
 
     public Bond(String ticker, String name, Currency currency,
-                double couponRate, int maturityYear) {
+                BigDecimal couponRate, int maturityYear) {
         super(ticker, name, currency);
         this.couponRate = couponRate;
         this.maturityYear = maturityYear;
     }
 
-    public double getCouponRate() {
+    public BigDecimal getCouponRate() {
         return couponRate;
     }
